// src/contracts/Abi.d.ts
export type AddressOutput = { value: string };

export type ContractIdInput = { value: string };

export type ContractIdOutput = { value: string };

export type ProjectInput = {
  project_id: BigNumberish;
  price: BigNumberish;
  max_buyers: BigNumberish;
  buyer_count: BigNumberish;
  owner_address: IdentityInput;
  metadata: string;
};

export type ProjectOutput = {
  project_id: BN;
  price: BN;
  max_buyers: BN;
  buyer_count: BN;
  owner_address: IdentityOutput;
  metadata: string;
};

export type VectorInput = {
  inner: [any, any, any, any, any];
  current_ix: BigNumberish;
};

export type VectorOutput = { inner: [any, any, any, any, any]; current_ix: BN };

export type IdentityInput = Enum<{
  Address: AddressInput;
  ContractId: ContractIdInput;
}>;

export type IdentityOutput = Enum<{
  Address: AddressOutput;
  ContractId: ContractIdOutput;
}>;